move task and user models to their own folder
