+++ client/src/app/models/task.model.ts
@@ -1,4 +1,4 @@
-export interface Task {
+export class Task {
     _id?: string;
     short_description: string;
     description: string;
