+++ client/src/app/services/user.service.ts
@@ -7,7 +7,7 @@ import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { catchError, map, tap } from 'rxjs/operators';
 
-import { User } from '../users/users'
+import { User } from '../models/user.model'
 import { MessageService } from '../services/message.service';
 
 const httpOptions = {
