--- client/src/app/tasks/tasks.ts
@@ -1,12 +1,0 @@
-export class Task {
-    _id?: string;
-    short_description: string;
-    description: string;
-    status: string;
-    category: string;
-    createdAt: string;
-    created_by: object;
-    updatedAt: string;
-    updated_by: object;
-    comments: Array<{}>;
-}
