+++ client/src/app/users/users.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { User } from './users';
+import { User } from '../models/user.model';
 import { UserService } from '../services/user.service';
 
 @Component({
